Update README.md
5/5 – (3 votes)

The purpose of this assignment is to practice the low-level concepts we have learned, from sequential logic to state machines in CircuitSim. Part one focuses on implementing sequential logic circuits like RS latches, Gated-D latches, D-Flip Flops whcih you will use to build a register from the ground up and eventually build a simple memory circuit as well. Part 2 focuses on using a state machine diagram and scenario to build and simplify KMaps and then build a Binary Reduced State Machine in CircuitSim. Finally, in Part 3, you will use your understanding of CircuitSim and these various circuitry concepts to build some of the key components of the Datapath: the PC, ALU, and Registers.

Objectives:

    To understand how a register circuit works

    To learn how to make a state machine

    To understand K-maps and simplification

    To see how state machines can integrate with computer processing

1.2 Task

There are four parts to this assignment.

    First, you will build sequential logic in CircuitSim from the ground up.

    Second, you will complete a K-Map to simplify your binary reduced state machine circuit, and imple-ment the simplified circuit.

    Third you will implement the datapath we have partially provided for you.

The assignment is due by 11:59 PM on February 8th, 2023. You could also submit the assignment by February 9th, 2023 for a late penalty of 25% of your overall grade. Before you start the assignment, be sure to read the rest of the PDF for more detailed instructions and hints.